requirements.txt
opencv-python==4.9.0.80
numpy==1.26.4
scikit-image==0.22.0
torch==2.2.1
torchvision==0.17.1
matplotlib==3.8.3
Pillow==10.2.0